Package 분리, Class 이름 변경
