+++ app/src/main/java/com/example/falldetectionapp/HomeActivity.java
@@ -15,7 +15,7 @@ import com.example.falldetectionapp.fragments.VideoFragment;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
 
-public class MainActivity extends AppCompatActivity {
+public class HomeActivity extends AppCompatActivity {
 
     private BottomNavigationView bottomNavigationView;
