+++ app/src/main/java/com/example/falldetectionapp/register/NokPhoneActivity.java
@@ -1,0 +1,16 @@
+package com.example.falldetectionapp.register;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.example.falldetectionapp.R;
+
+public class NokPhoneActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_emergency_phone);
+    }
+}
