+++ app/src/main/res/layout/activity_emergency_phone.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".emergency_phone">
+    tools:context=".register.NokPhoneActivity">
 
     <TextView
         android:layout_width="wrap_content"
