+++ app/src/main/res/layout/activity_log_in.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".log_in">
+    tools:context=".LoginActivity">
 
     <ImageView
         android:id="@+id/image_login"
