+++ app/src/main/res/layout/activity_sign_up.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:foregroundTint="#FEFEFE"
     android:orientation="vertical"
-    tools:context=".sign_up">
+    tools:context=".register.RegisterActivity">
 
     <TextView
         android:layout_width="match_parent"
